# Use `torch.where` for faster contract